Rename test helpers folder
