+++ packages/domain-search/jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
 	preset: '../../test/packages/jest-preset.js',
 	testEnvironment: 'jsdom',
-	setupFilesAfterEnv: [ '<rootDir>/src/test/setup.ts' ],
+	setupFilesAfterEnv: [ '<rootDir>/src/test-helpers/setup.ts' ],
 };
