+++ packages/domain-search/src/test-helpers/factories/availability.ts
