+++ packages/domain-search/src/test-helpers/factories/suggestions.ts
