+++ packages/domain-search/src/test-helpers/mocks/availability.ts
