+++ packages/domain-search/src/test-helpers/mocks/suggestions.ts
